requirements.txt
PGPy==0.3.0
PyHawk-with-a-single-extra-commit==0.1.5
PyYAML==3.11
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
cffi==0.9.2
cryptography==0.6
enum34==1.0.4
kombu==3.0.24
pycparser==2.10
pystache==0.5.4
requests==2.6.0
singledispatch==3.4.0.3
six==1.9.0
taskcluster==0.0.16
wsgiref==0.1.2
